+++ utils/src/retriever.c
@@ -12,32 +12,32 @@ retriever_t *retriever_init(decomp_t *dp, int period) {
   retriever_t *rp = (retriever_t *) malloc(1 * sizeof(retriever_t));
   rp->period = period;
   rp->current_step = 0;
-  rp->data = (double *) malloc(period * dp->size * sizeof(double));
+  rp->data = (float *) malloc(period * dp->size * sizeof(float));
   rp->dp = dp;
   return rp;
 }
 
-void retriever_feed(retriever_t *rp, double *I) {
+void retriever_feed(retriever_t *rp, float *I) {
   // Store data so that they can be queried
   // Data can be compressed or reorganized here
-  memcpy(rp->data + rp->dp->size*rp->current_step, I, rp->dp->size*sizeof(double));
+  memcpy(rp->data + rp->dp->size*rp->current_step, I, rp->dp->size*sizeof(float));
   // Update step
   rp->current_step = (rp->current_step + 1) % rp->period;
 }
 
 // Return pointer to that step data
-double *retriever_get_step(retriever_t *rp, int step) {
+float *retriever_get_step(retriever_t *rp, int step) {
   return rp->data + rp->dp->size*step;
 }
 
-double *retriever_get_laststep(retriever_t *rp) {
+float *retriever_get_laststep(retriever_t *rp) {
   int step;
   step = rp->current_step == 0 ? rp->period-1 : rp->current_step-1;
   return retriever_get_step(rp, step);
 }
 
 // Return size
-int retriever_get_chunk(retriever_t *rp, int pos, double *data) {
+int retriever_get_chunk(retriever_t *rp, int pos, float *data) {
   int i, j, k;
   int lrow, lcol, orow, ocol;
 
@@ -53,7 +53,7 @@ int retriever_get_chunk(retriever_t *rp, int pos, double *data) {
   return rp->period * lrow * lcol;
 }
 
-/* void retriever_get(retriever_t *rp, int *pos, int pos_count, double *data) { */
+/* void retriever_get(retriever_t *rp, int *pos, int pos_count, float *data) { */
 /*   int i; */
 /*   for (i = 0; i < pos_count; i++) { */
 /*     retriever_get_chunk(pos[i], data+i*rp->period); */
